refactor(server): route naming
