+++ server/src/api/admin/role/:role.ts
@@ -11,6 +11,27 @@ import { adminSchemas } from "lib";
 const router = express.Router();
 
 router.get(
+    "/",
+    checkAccess(),
+    asyncHandler(async (req, res, _) => {
+        const { q: searchQuery } = adminSchemas.roleSearchQuerySchema.parse(
+            req.query
+        );
+        const allRoles = await db.query.roles.findMany({
+            columns: {
+                roleName: true,
+                memberCount: true,
+            },
+            where: (fields, { ilike }) =>
+                searchQuery?.length
+                    ? ilike(fields.roleName, `%${searchQuery}%`)
+                    : undefined,
+        });
+        res.json(allRoles);
+    })
+);
+
+router.get(
     "/:role",
     checkAccess(),
     asyncHandler(async (req, res, next) => {
